neuconn_app/requirements.txt
# NeuConn Dependencies

# Core
streamlit>=1.30.0
pyyaml>=6.0.1
jsonschema>=4.20.0

# Neuroimaging
nibabel>=5.3.0
nilearn>=0.13.0
pybids>=0.16.0

# Data handling
pandas>=2.3.0
numpy>=1.26.0

# Scientific computing
scipy>=1.16.0
statsmodels>=0.14.5

# Visualization
matplotlib>=3.8.0
plotly>=6.3.0
seaborn>=0.13.0

# Network analysis
networkx>=3.2

# Time series analysis
nitime>=0.10  # For MVAR/Granger

# HPC connection
paramiko>=3.4.0

# Templates
Jinja2>=3.1.3

# Web components
streamlit-aggrid>=0.3.4  # For interactive tables (optional)
